# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
      "/manage/collection/create": {
        "post": {
          "description": "创建模型集合",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "ModelCollection"
          ],
          "summary": "创建模型集合",
          "parameters": [
            {
              "description": "模型集合",
              "name": "model",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/schema.ModelCollection"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功创建的模型集合",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-schema_ModelCollection"
              }
            }
          }
        }
      },
        "/manage/collection/list": {
            "get": {
                "description": "批量获取模型集合",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelCollection"
                ],
                "summary": "批量获取模型集合",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_expr",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "模型集合列表",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-entity_PaginatedTotalResponse-schema_ModelCollection"
                        }
                    }
                }
            }
        },
      "/manage/collection/{id}": {
            "get": {
                "description": "获取模型集合",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModelCollection"
                ],
                "summary": "获取模型集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ModelCollection ID",
                        "name": "collection_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "模型",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_ModelCollection"
                        }
                    }
                }
            }
        },
      "/manage/collection/{id}/delete": {
        "post": {
          "description": "删除模型集合",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Model"
          ],
          "summary": "删除模型集合",
          "parameters": [
            {
              "type": "integer",
              "description": "ModelCollection ID",
              "name": "collection_id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "删除成功与否",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-bool"
              }
            }
          }
        }
      },
      "/manage/collection/{id}/update": {
        "post": {
          "description": "更新模型集合",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "ModelCollection"
          ],
          "summary": "更新模型集合",
          "parameters": [
            {
              "type": "integer",
              "description": "模型集合 ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "模型集合更新参数",
              "name": "model",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/entity.ReqUpdateBody-schema_ModelCollection"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "成功更新与否",
              "schema": {
                "$ref": "#/definitions/entity.CommonResponse-bool"
              }
            }
          }
        }
      },
...
        "/manage/model/create": {
            "post": {
                "description": "创建模型并绑定到 API 供应商",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "创建模型",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功创建的模型",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Model"
                        }
                    }
                }
            }
        },
        "/manage/model/delete/{model_id}": {
            "post": {
                "description": "删除模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "删除模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Model ID",
                        "name": "model_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功与否",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-bool"
                        }
                    }
                }
            }
        },
        "/manage/model/list": {
            "get": {
                "description": "批量获取模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "批量获取模型",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_expr",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "模型列表",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-entity_PaginatedTotalResponse-schema_Model"
                        }
                    }
                }
            }
        },
        "/manage/model/provider/{provider_id}": {
            "get": {
                "description": "获取 API 提供商的模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取 API 提供商的模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Provider ID",
                        "name": "provider_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_expr",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "模型列表",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-entity_PaginatedTotalResponse-schema_Model"
                        }
                    }
                }
            }
        },
...
        "/manage/model/update": {
            "post": {
                "description": "Update 更新模型，若参数不传入或为空，则不会更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "更新模型",
                "parameters": [
                    {
                        "description": "模型",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功与否",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-bool"
                        }
                    }
                }
            }
        },
        "/manage/model/{model_id}": {
            "get": {
                "description": "获取模型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Model"
                ],
                "summary": "获取模型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Model ID",
                        "name": "model_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "模型",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Model"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CommonResponse-bool": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "type": "boolean"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-entity_PaginatedTotalResponse-schema_Model": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PaginatedTotalResponse-schema_Model"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
        "entity.CommonResponse-entity_PaginatedTotalResponse-schema_ModelCollection": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PaginatedTotalResponse-schema_ModelCollection"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-schema_Model": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Model"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
        "entity.CommonResponse-schema_ModelCollection": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ModelCollection"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.PaginatedTotalResponse-schema_Model": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Model"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entity.PaginatedTotalResponse-schema_ModelCollection": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ModelCollection"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
      "entity.ReqUpdateBody-schema_ModelCollection": {
        "type": "object",
        "required": [
          "data",
          "updates"
        ],
        "properties": {
          "data": {
            "$ref": "#/definitions/schema.ModelCollection"
          },
          "updates": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          }
        }
      },
...
        "schema.APIKey": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "description": "API 密钥",
                    "type": "string"
                },
                "provider_id": {
                    "description": "外键，指向 Provider",
                    "type": "integer"
                }
            }
        },
...
        "schema.Model": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "使用 JSON 储存配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ModelConfig"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "额外模型描述",
                    "type": "string"
                },
                "display_name": {
                    "description": "对外展示模型名称",
                    "type": "string"
                },
                "id": {
                    "description": "原始数据",
                    "type": "integer"
                },
                "name": {
                    "description": "模型名称",
                    "type": "string"
                },
                "provider": {
                    "description": "组装数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Provider"
                        }
                    ]
                },
                "provider_id": {
                    "description": "关联的 Provider ID",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "schema.ModelCollection": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "额外描述",
                    "type": "string"
                },
                "display_name": {
                    "description": "展示名称",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "models": {
                    "description": "关联的模型",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Model"
                    }
                },
                "name": {
                    "description": "唯一标识名称",
                    "type": "string"
                }
            }
        },
        "schema.ModelConfig": {
            "type": "object",
            "properties": {
                "allow_system_prompt": {
                    "description": "是否允许用户自行修改系统提示",
                    "type": "boolean"
                },
                "default_temperature": {
                    "description": "默认温度",
                    "type": "number"
                },
                "frequency_penalty": {
                    "type": "number"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "presence_penalty": {
                    "type": "number"
                },
                "system_prompt": {
                    "description": "预设系统提示",
                    "type": "string"
                },
                "top_p": {
                    "type": "number"
                }
            }
        },
...
        "schema.Provider": {
            "type": "object",
            "properties": {
                "api_keys": {
                    "description": "一对多关系，与 APIKey 模型关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.APIKey"
                    }
                },
                "base_url": {
                    "description": "API 的基本 URL",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "额外提供商描述",
                    "type": "string"
                },
                "display_name": {
                    "description": "对外展示提供商名称",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "models": {
                    "description": "一对多关系，与 Model 模型关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Model"
                    }
                },
                "name": {
                    "description": "提供商名称",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}